+++ lib/redis-objects-daily-counter.rb
@@ -14,7 +14,7 @@ class Redis
   module Objects
     PERIODICALS.each do |periodical|
       autoload :"#{periodical.capitalize}Counters", "redis/objects/periodical_counters"
-      autoload :"#{periodical.capitalize}Hashes", "redis/objects/#{periodical}_hashes"
+      autoload :"#{periodical.capitalize}Hashes", "redis/objects/periodical_hashes"
       autoload :"#{periodical.capitalize}Sets", "redis/objects/#{periodical}_sets"
     end
